Move `NativeAllocationTracer` class to `tools` package
